🚨 Add simple test for list
For now only tests that the functions list is empty (which we expect it
to be given our current implementation)
