+++ protocols/functions.proto
@@ -4,7 +4,7 @@ syntax = "proto3";
 // effectively a namespace
 package functions;
 
-service FunctionsService {
+service Functions {
   rpc List (ListRequest) returns (ListResponse);
 }
