+++ services/avery/Cargo.toml
@@ -16,5 +16,8 @@ tonic = "0.1"
 url = "2"
 uuid = { version = "0.8", features = ["serde", "v4"] }
 
+[dev-dependencies]
+futures = "0.3"
+
 [build-dependencies]
 tonic-build = "0.1"
